email service updated

+++ client/src/components/admin/SubCategory/SubCategory.jsx
@@ -133,13 +133,15 @@ class SubCategory extends Component {
 
     // handle sub category deletion
     handleDelete = (_id) => {
-        deleteSubCategories(_id).then(res => {
-            getSubCategories().then(res => {
-                this.setState({
-                    subCategories: res.data
+        if (window.confirm("Do you need to remove this category? This category may contain some products")) {
+            deleteSubCategories(_id).then(res => {
+                getSubCategories().then(res => {
+                    this.setState({
+                        subCategories: res.data
+                    })
                 })
-            })
-        });
+            });
+        }
     }
 
     render() {
